+++ src/lowrank/batchsvd.h
@@ -1,6 +1,6 @@
 
 #include <complex.h>
 
-extern void batch_svthresh(long M, long N, long num_blocks, float lambda, complex float dst[num_blocks][N][M]);
+extern void batch_svthresh(int M, int N, int num_blocks, float lambda, complex float dst[num_blocks][N][M]);
